--- src/_base.scss
@@ -1,28 +1,0 @@
-%center {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  %border-box {
-    box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    -webkit-box-sizing: border-box;
-  }
-  
-  %slide-in {
-    animation: slide-up 1s ease-out forwards;
-  }
-  %slide-out {
-    animation: slide-down 1s ease-out forwards;
-  }
-  
-  %fade-in {
-    opacity: 1 !important;
-    transition: opacity 1s;
-  }
-  
-  %fade-out {
-    opacity: 0;
-    transition: opacity 1s;
-  }
